+++ cpp/arduino/Arduino.cpp
@@ -9,15 +9,15 @@ void digitalWrite(unsigned char pin, unsigned char val) {
 
 int digitalRead(unsigned char pin) {
   GodmodeState* godmode = GODMODE();
-  return godmode->digitalPin[pin];
-}
-
-int analogRead(unsigned char pin) {
-  GodmodeState* godmode = GODMODE();
-  return godmode->analogPin[pin];
+  return godmode->digitalPin[pin].retrieve();
 }
 
 void analogWrite(unsigned char pin, int val) {
   GodmodeState* godmode = GODMODE();
   godmode->analogPin[pin] = val;
+}
+
+int analogRead(unsigned char pin) {
+  GodmodeState* godmode = GODMODE();
+  return godmode->analogPin[pin].retrieve();
 }